unit test folder housekeeping
