+++ src/Dhgms.GripeWithRoslyn.UnitTests/Analyzers/Language/UseTypeofInsteadOfBaseMethodDeclaringTypeAnalyzerTest.cs
@@ -2,8 +2,8 @@
 // This file is licensed to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
+using System.Reflection;
 using Dhgms.GripeWithRoslyn.Analyzer;
-using Dhgms.GripeWithRoslyn.Analyzer.Analyzers;
 using Dhgms.GripeWithRoslyn.Analyzer.Analyzers.Language;
 using Dhgms.GripeWithRoslyn.UnitTests.Helpers;
 using Microsoft.CodeAnalysis;
@@ -11,12 +11,12 @@ using Microsoft.CodeAnalysis.Diagnostics;
 using Xunit;
 using CodeFixVerifier = Dhgms.GripeWithRoslyn.UnitTests.Verifiers.CodeFixVerifier;
 
-namespace Dhgms.GripeWithRoslyn.UnitTests.Analyzers
+namespace Dhgms.GripeWithRoslyn.UnitTests.Analyzers.Language
 {
     /// <summary>
-    /// Unit Tests for <see cref="UseTypeofInsteadOfTypeGetTypeAnalyzer"/>.
+    /// Unit Tests for <see cref="UseTypeofInsteadOfBaseMethodDeclaringTypeAnalyzer"/>.
     /// </summary>
-    public sealed class UseTypeofInsteadOfTypeGetTypeAnalyzerTest : CodeFixVerifier
+    public sealed class UseTypeofInsteadOfBaseMethodDeclaringTypeAnalyzerTest : CodeFixVerifier
     {
         /// <summary>
         /// Test to ensure bad code returns a warning.
@@ -33,14 +33,17 @@ namespace Dhgms.GripeWithRoslyn.UnitTests.Analyzers
         {
             public void MethodName()
             {
-                System.Type.GetType(""System.String"");
+                global::System.Reflection.MethodBase.GetCurrentMethod().DeclaringType;
             }
         }
     }";
+
+            var a = MethodBase.GetCurrentMethod().DeclaringType;
+
             var expected = new DiagnosticResult
             {
                 Id = DiagnosticIdsHelper.UseEncodingUnicodeInsteadOfAscii,
-                Message = "Consider usage of typeof(x) instead of System.Type.GetType.",
+                Message = "Consider usage of typeof(x) instead of MethodBase.GetCurrentMethod().DeclaringType.",
                 Severity = DiagnosticSeverity.Warning,
                 Locations =
                     new[]
@@ -55,7 +58,7 @@ namespace Dhgms.GripeWithRoslyn.UnitTests.Analyzers
         /// <inheritdoc />
         protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
         {
-            return new UseTypeofInsteadOfTypeGetTypeAnalyzer();
+            return new UseTypeofInsteadOfBaseMethodDeclaringTypeAnalyzer();
         }
     }
 }
