+++ src/Dhgms.GripeWithRoslyn.UnitTests/Analyzers/Language/UseTypeofInsteadOfTypeGetTypeAnalyzerTest.cs
@@ -2,9 +2,7 @@
 // This file is licensed to you under the MIT license.
 // See the LICENSE file in the project root for full license information.
 
-using System.Reflection;
 using Dhgms.GripeWithRoslyn.Analyzer;
-using Dhgms.GripeWithRoslyn.Analyzer.Analyzers;
 using Dhgms.GripeWithRoslyn.Analyzer.Analyzers.Language;
 using Dhgms.GripeWithRoslyn.UnitTests.Helpers;
 using Microsoft.CodeAnalysis;
@@ -12,12 +10,12 @@ using Microsoft.CodeAnalysis.Diagnostics;
 using Xunit;
 using CodeFixVerifier = Dhgms.GripeWithRoslyn.UnitTests.Verifiers.CodeFixVerifier;
 
-namespace Dhgms.GripeWithRoslyn.UnitTests.Analyzers
+namespace Dhgms.GripeWithRoslyn.UnitTests.Analyzers.Language
 {
     /// <summary>
-    /// Unit Tests for <see cref="BooleanTryMethodShouldBeUsedInLogicalNotIfStatementAnalyzer"/>.
+    /// Unit Tests for <see cref="UseTypeofInsteadOfTypeGetTypeAnalyzer"/>.
     /// </summary>
-    public sealed class BooleanTryMethodShouldBeUsedInLogicalNotIfStatementAnalyzerTests : CodeFixVerifier
+    public sealed class UseTypeofInsteadOfTypeGetTypeAnalyzerTest : CodeFixVerifier
     {
         /// <summary>
         /// Test to ensure bad code returns a warning.
@@ -34,15 +32,14 @@ namespace Dhgms.GripeWithRoslyn.UnitTests.Analyzers
         {
             public void MethodName()
             {
-                int.TryParse(""x"", out var result);
+                System.Type.GetType(""System.String"");
             }
         }
     }";
-
             var expected = new DiagnosticResult
             {
-                Id = DiagnosticIdsHelper.TryParseShouldBeUsedInLogicalNotIfStatement,
-                Message = BooleanTryMethodShouldBeUsedInLogicalNotIfStatementAnalyzer.Title,
+                Id = DiagnosticIdsHelper.UseEncodingUnicodeInsteadOfAscii,
+                Message = "Consider usage of typeof(x) instead of System.Type.GetType.",
                 Severity = DiagnosticSeverity.Warning,
                 Locations =
                     new[]
@@ -57,7 +54,7 @@ namespace Dhgms.GripeWithRoslyn.UnitTests.Analyzers
         /// <inheritdoc />
         protected override DiagnosticAnalyzer GetCSharpDiagnosticAnalyzer()
         {
-            return new BooleanTryMethodShouldBeUsedInLogicalNotIfStatementAnalyzer();
+            return new UseTypeofInsteadOfTypeGetTypeAnalyzer();
         }
     }
 }
